HOTFIX-EOI-RFI-CSVCOUMN-UPDATE
